[dharam]  implemented mailer user and mailer module
What changes have you made?

implemented user module to create a new user, later which will be act as
article writer and he can write, edit and delete an article related to him.
a user which is not registered with us will not be able create any article
he  can only read articles.
	Also implemented a mailer module which will send a welcome e mail to
user on successful registration. and later we will use this module to help
the user to recover his password