[dharam] resolved a bug
What changes have you made?
There was a bug in blog controller
when a user is logged in and he is on home page, and he click on the 'more'
link provided to the particular blog he will get a error message that the blogs
is undefined for nill user class.
So what was happing there is on the home page the link url to display the full blog
dis not have any reference to user_id and in blog controller we were checking that
if user is logged in we first find the user and then his blog so while fetching the
user from data base we were using user_id to get the user which was not there in
params object so the user was nil
 so now along with logged_in? check I am checking wether the params have user_id or
not if yes themn we do first find user then blog but if not then we will directly find
the blog using blog id which is already there in params object